# internal/types/types.go
// Code generated by goctl. DO NOT EDIT.
// goctl 1.7.5

package types

type Base struct {
	Code int    `json:"code"`
	Msg  string `json:"msg"`
}

type ConsumeCheckDelinquencyReq struct {
	LoanID int64 `json:"loanId"`
}

type ConsumeCheckDelinquencyResp struct {
	Base
}

type ConsumeDelinquencyReq struct {
	LoanID int64 `json:"loanId"`
}

type ConsumeDelinquencyResp struct {
	Base
}

type ConsumeLateFeeReq struct {
	LoanID int64 `json:"loanId"`
}

type ConsumeLateFeeResp struct {
	Base
}

type ConsumeRepaymentReq struct {
	PaymentID int64 `json:"paymentId"`
}

type ConsumeRepaymentResp struct {
	Base
	LoanID int64 `json:"loanId"`
}

type CreateLoanReq struct {
	PrincipalAmount int64 `json:"principalAmount"`
	TermWeeks       int64 `json:"termWeeks"`
}

type CreateLoanResp struct {
	Base
	LoanID int64 `json:"loanId"`
}

type CreateRepaymentReq struct {
	LoanID        int64   `path:"loanId"`
	PaymentAmount float64 `json:"paymentAmount"`
}

type CreateRepaymentResp struct {
	Base
	PaymentIDs []*DataCreateRepayment `json:"paymentIds"`
}

type CronDelinquencyReq struct {
	QueryLimit int `form:"queryLimit,default=1000"`
}

type CronDelinquencyResp struct {
	Base
	Data []*DataDelinquency `json:"data"`
}

type CronLateFeeReq struct {
	QueryLimit int `form:"queryLimit,default=1000"`
}

type CronLateFeeResp struct {
	Base
	Data []*DataLoanScheduleLate `json:"data"`
}

type DataCreateRepayment struct {
	PaymentId int64 `json:"paymentId"`
}

type DataDelinquency struct {
	LoanID int64 `json:"loanId"`
}

type DataLoanScheduleLate struct {
	LoanID int64 `json:"loanId"`
}

type GetLoanDelinquencyReq struct {
	LoanID int64 `path:"loanId"`
}

type GetLoanDelinquencyResp struct {
	Base
	LoanID       int64 `json:"loanId"`
	IsDelinquent bool  `json:"isDelinquent"`
}

type GetLoanOutStandingResp struct {
	Base
	LoanID             int64 `json:"loanId"`
	OutstandingBalance int64 `json:"outstandingBalance"`
}

type GetLoanOutstandingReq struct {
	LoanID int64 `path:"loanId"`
}

type GetRepaymentReq struct {
	PaymentID int64 `path:"paymentId"`
}

type GetRepaymentResp struct {
	Base
	PaymentID     int64  `json:"paymentId"`
	PaymentAmount int64  `json:"paymentAmount"`
	PaymentDate   string `json:"paymentDate"`
	Status        string `json:"status"`
}