# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "billing API",
    "description": "simple billing API",
    "version": "1.0.0"
  },
  "host": "127.0.0.1:8888",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/billing/delinquency/{loanId}": {
      "get": {
        "operationId": "GetLoanDelinquencyHandler",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetLoanDelinquencygResp"
            }
          }
        },
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "loans"
        ]
      }
    },
...
    "/v1/billing/pay/{loanId}": {
      "post": {
        "operationId": "CreateRepaymentHandler",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateRepaymentResp"
            }
          }
        },
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateRepaymentReq"
            }
          }
        ],
        "tags": [
          "loans"
        ]
      }
    },
...
  },
  "definitions": {
...
    "CreateRepaymentReq": {
      "type": "object",
      "properties": {
        "paymentAmount": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "CreateRepaymentReq",
      "required": [
        "paymentAmount"
      ]
    },
    "CreateRepaymentResp": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "paymentIds": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DataCreateRepayment"
          }
        }
      },
      "title": "CreateRepaymentResp",
      "required": [
        "paymentIds"
      ]
    },
...
    "DataCreateRepayment": {
      "type": "object",
      "properties": {
        "paymentId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "DataCreateRepayment",
      "required": [
        "paymentId"
      ]
    },
...
    "GetLoanDelinquencygResp": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "loanId": {
          "type": "integer",
          "format": "int64"
        },
        "isDelinquent": {
          "type": "boolean",
          "format": "boolean"
        }
      },
      "title": "GetLoanDelinquencygResp",
      "required": [
        "loanId",
        "isDelinquent"
      ]
    },
...
  },
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "description": "Enter JWT Bearer token **_only_**",
      "name": "Authorization",
      "in": "header"
    }
  }
}